Load config by default
Users that require "google/cloud" should be able to access the configuration
without first requiring "google/cloud/config", which is an internal path.
